finish file alignment
