+++ config/splat.pspeu.dra.yaml
@@ -50,8 +50,6 @@ segments:
       - [0x60F8, c, dra/628AC] # O4
       - [0x66B0, c, dra/5F60C] # O0
       - [0xA710, c, dra_psp/A710] # O4
-      - [0xBDE0, c, dra_psp/BDE0]
-      - [0xC0B0, c, dra_psp/C0B0] # O4
       - [0xE6A8, c, dra_psp/E6A8]
       - [0x18210, c, dra_psp/18210]
       - [0x186E8, c, dra_psp/186E8]
@@ -143,7 +141,6 @@ segments:
       - [0xEE3C0, .rodata, dra/628AC]
       - [0xEE3E8, .rodata, dra/628AC]
       - [0xEE408, .rodata, dra_psp/A710]
-      - [0xEE428, .rodata, dra_psp/C0B0]
       - [0xEE4C8, .rodata, dra_psp/E6A8]
       - [0xEE570, .rodata, dra_psp/19AB0]
       - [0xEE590, .rodata, dra_psp/1A1A0]
