--- src/dra_psp/BDE0.c
@@ -1,8 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../dra/dra.h"
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/BDE0", func_psp_090E8760);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/BDE0", func_800F223C);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/BDE0", func_psp_090E88C0);
